Support fractional silence thresholds + add a talk-to key logger
Two fixes for on-hardware issues found testing streaming:

1. Quiet-mic support: the sox silence threshold (vad_threshold_pct,
   [streaming].silence_threshold_pct) is now an f32, so a mic whose speech
   peaks below 1% (e.g. an audio interface at low gain) can set a fractional
   percent like 0.3 instead of being trimmed to "no speech". A lenient
   deserializer accepts both a bare integer (3) and a float (0.3), so existing
   configs keep parsing; the sox arg renders 0.3 -> "0.3%".

2. Key diagnostics: GHOSTTY_VOICE_KEYLOG=1 talk-to <cmd> logs raw stdin bytes
   in caret notation to the talk-to log, to capture the exact escape sequence a
   wrapped TUI's keyboard mode produces for a trigger key (so triggers that work
   under `cat -v` but not under a TUI that changes keyboard mode can be fixed).

config.toml.example documents both. Tests cover integer+fractional parsing and
the 0.3% sox-arg render.

+++ crates/ghostty-voice-io/src/audio.rs
@@ -39,7 +39,7 @@ pub fn spawn_vad_recorder(
     device: &str,
     out: &Path,
     silence_seconds: f32,
-    threshold_pct: u32,
+    threshold_pct: f32,
 ) -> Result<Child> {
     spawn_silence_stopped_recorder(device, out, silence_seconds, threshold_pct)
 }
@@ -55,7 +55,7 @@ pub fn spawn_streaming_recorder(
     device: &str,
     out: &Path,
     session_end_silence_seconds: f32,
-    threshold_pct: u32,
+    threshold_pct: f32,
 ) -> Result<Child> {
     spawn_silence_stopped_recorder(device, out, session_end_silence_seconds, threshold_pct)
 }
@@ -68,7 +68,7 @@ fn spawn_silence_stopped_recorder(
     device: &str,
     out: &Path,
     silence_seconds: f32,
-    threshold_pct: u32,
+    threshold_pct: f32,
 ) -> Result<Child> {
     let argv = ghostty_voice_core::vad::record_args(
         &out.to_string_lossy(),
@@ -98,7 +98,7 @@ pub fn spawn_continuous_recorder(
     device: &str,
     out_template: &Path,
     clip_pause_seconds: f32,
-    threshold_pct: u32,
+    threshold_pct: f32,
 ) -> Result<Child> {
     let argv = ghostty_voice_core::vad::continuous_record_args(
         &out_template.to_string_lossy(),
@@ -382,7 +382,7 @@ mod tests {
         assert!(full >= Duration::from_secs(3), "source should be ~3.5 s");
 
         // Apply the same trailing-silence trim the VAD recorder uses (2 s @ 3%).
-        let effect = ghostty_voice_core::vad::silence_effect(2.0, 3);
+        let effect = ghostty_voice_core::vad::silence_effect(2.0, 3.0);
         let status = Command::new("sox")
             .arg(&src)
             .arg(&out)
@@ -436,7 +436,7 @@ mod tests {
             .expect("sox synth silence");
         assert!(synth.success(), "sox synth failed");
 
-        let effect = ghostty_voice_core::vad::silence_effect(2.0, 3);
+        let effect = ghostty_voice_core::vad::silence_effect(2.0, 3.0);
         let status = Command::new("sox")
             .arg(&src)
             .arg(&out)
@@ -500,7 +500,7 @@ mod tests {
         // The exact continuous split effect: cut after 1 s of silence @ 3%,
         // newfile + restart. Output template uses %n for the clip index.
         let template = dir.join("clip-%n.wav");
-        let effect = ghostty_voice_core::vad::continuous_split_effect(1.0, 3);
+        let effect = ghostty_voice_core::vad::continuous_split_effect(1.0, 3.0);
         let status = Command::new("sox")
             .arg(&src)
             .arg(&template)
